fixed warning on depthImageBasedGeometry

+++ core/base/depthImageBasedGeometryApproximation/DepthImageBasedGeometryApproximation.h
@@ -250,7 +250,7 @@ int ttk::DepthImageBasedGeometryApproximation::execute(
 #ifdef TTK_ENABLE_OPENMP
 #pragma omp parallel for num_threads(this->threadNumber_)
 #endif
-    for(int i = 0; i < nbTri; i++) {
+    for(size_t i = 0; i < nbTri; i++) {
       cells_off[i] = i * 3;
     }
